Consolidate 'player' package between games.strategy.common and games.strategy.triplea
