+++ src/games/strategy/triplea/TripleAPlayer.java
@@ -13,7 +13,7 @@ import javax.swing.AbstractAction;
 import javax.swing.ButtonModel;
 import javax.swing.SwingUtilities;
 
-import games.strategy.common.player.AbstractHumanPlayer;
+import games.strategy.triplea.player.AbstractHumanPlayer;
 import games.strategy.swing.SwingAction;
 import games.strategy.debug.ClientLogger;
 import games.strategy.engine.data.GameData;
