+++ src/games/strategy/triplea/ai/AbstractAI.java
@@ -13,7 +13,7 @@ import java.util.Map.Entry;
 import java.util.Set;
 import java.util.logging.Logger;
 
-import games.strategy.common.player.AbstractBaseAI;
+import games.strategy.triplea.player.AbstractBaseAI;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.data.Resource;
