+++ src/games/strategy/triplea/player/AbstractBasePlayer.java
@@ -1,4 +1,4 @@
-package games.strategy.common.player;
+package games.strategy.triplea.player;
 
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
