+++ src/games/strategy/triplea/player/ai/GameState.java
@@ -1,4 +1,4 @@
-package games.strategy.common.player.ai;
+package games.strategy.triplea.player.ai;
 
 import java.util.Collection;
